Split the project into separate library and binary components to allow for (circular) dependencies across binaries  - Binaries are shallow and interface to duplicate library files  - lib.rs is the only place `mod` gets called  - All other imports are handled with `use crate::xxxx` (in lib files) or `use veristone:xxxx` (in binaries)  - Refactored some of read_blif and mcfunction to move mcfunction calls up to main & set up for routing in mcfunction
