+++ src/main.rs
@@ -1,14 +1,9 @@
 use std::path::Path;
 use clap::Parser;
 
-#[path = "./bin/make_blif.rs"]
-mod make_blif;
-
-#[path = "./bin/read_blif.rs"]
-mod read_blif;
-
-#[path = "./bin/mcfunction.rs"]
-mod mcfunction;
+use veristone::make_blif;
+use veristone::read_blif;
+use veristone::mcfunction;
 
 #[derive(Parser, Debug)]
 #[command(version, about, long_about = None)]
@@ -23,5 +18,6 @@ fn main() {
     let blif_buf = make_blif::make_blif_path(verilog_path);
     let blif_path = blif_buf.as_path();
     make_blif::generate_blif(verilog_path, Path::new("res/mc.lib"), blif_path, false);
-    read_blif::read_blif(blif_path);
+    let (gates, wires) = read_blif::read_blif(blif_path, read_blif::PlacementAlgo::DumbGrid { num_cols: 4 });
+    mcfunction::write_mcfunction(&gates, &wires, mcfunction::RoutingAlgo::Wireless);
 }
